requirements.txt
# Core dependencies
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.21.0,<1.0.0
numpy>=1.25.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.12.0,<1.0.0
joblib>=1.3.0,<2.0.0
pydantic>=2.7.4,<3.0.0
torch>=2.0.0,<3.0.0
reportlab>=4.0.0,<5.0.0

# LangChain + LangGraph for AI Assistant
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# Configuration & CLI
python-dotenv>=1.0.0,<2.0.0
typer>=0.9.0,<1.0.0

# Development & Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
mypy>=1.0.0,<2.0.0
flake8>=6.0.0,<7.0.0
black>=23.0.0,<25.0.0

# Optional utilities
tqdm>=4.65.0,<5.0.0
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<1.0.0
